backend-python/requirements.txt
# Python Backend Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: Use python-adb instead of subprocess (more reliable)
# adb-shell>=0.4.0
# pure-python-adb>=0.3.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# pylint>=2.17.0
